Enable MC scaling per region in plot_data_mc.py
